+++ docs/source/sources/changelog.rst
@@ -3,6 +3,10 @@
 Changelog
 ===========
 
+**Version 1.2.2**
+ Added new adaptive filters: :ref:`filter-vslms_mathews`,
+ :ref:`filter-vslms_benveniste`, :ref:`filter-vslms_ang`.
+
 **Version 1.2.1** *Released: 2022-02-07*
  Bugfix of the main adaptive filter class.